requirements.txt
customtkinter
qrcode
opencv-python
pillow